svg-precision-skill/scripts/requirements.txt
# Optional dependencies for preview + diff
cairosvg>=2.5
pillow>=9
